requirements.txt
absl-py==1.0.0
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.10
Django==4.0.1
djangorestframework==3.13.1
idna==3.3
numpy==1.22.0
ortools==9.2.9972
pika==1.2.0
pip==21.1.2
protobuf==3.19.3
pytz==2021.3
requests==2.27.1
setuptools==57.0.0
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8
wheel==0.36.2